requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==3.0.5
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-eventstream==5.3.1
django-filter==24.3
django-grip==3.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gripcontrol==4.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
MarkupSafe==3.0.2
pillow==11.0.0
pubcontrol==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
zope.interface==7.2
